fix: move PauseDuration + computePausedUntil out of 'use server' module
Next 16 requires every export from a 'use server' file to be async
(server actions). computePausedUntil is a sync utility, so Turbopack
rejects it. Move it to a plain module (pause-duration.ts) and
re-export the type from buyer-actions for backward compat.
